Fixes file append to act sanely. Fixes ip_src and ip_dst assignments to work with Numerics, Octets, and Strings.
git-svn-id: https://packetfu.googlecode.com/svn/trunk@192 ebf01b17-4b54-0410-ae1b-bdaaa366248d

+++ lib/packetfu/pcap.rb
@@ -451,7 +451,11 @@ module PacketFu
 			end
 			append = args[:append]
 			if append
-				File.open(filename,'ab') {|file| file.write(self.body.to_s)}
+				if File.exists? filename
+					File.open(filename,'ab') {|file| file.write(self.body.to_s)}
+				else
+					File.open(filename,'wb') {|file| file.write(self.to_s)}
+				end
 			else
 				File.open(filename,'wb') {|file| file.write(self.to_s)}
 			end
